// src/app/shared/models/Product.ts
export class Product{

    brandName:String;
    productName: String;
    weight: String;
    price: number;
    mrp: number;
    imageUrl: String;
    offerText: number;
    quantity: number;

    // constructor(){
    //     this.quantity = 1;
    // }

    // constructor(brandName,productName,quantity,price,mrp,imageUrl,offerText){
    //     this.brandName=brandName;
    //     this.productName=productName;
    //     this.quantity=quantity;
    //     this.price=price;
    //     this.mrp=mrp;
    //     this.imageUrl=imageUrl;
    //     this.offerText=offerText;
    // }
}